Provide shared default settings and test
