+++ client/test/defaultSettingsLoad.test.ts
@@ -1,0 +1,14 @@
+import storage, { getItemSync } from '../src/storage';
+import { defaultSettings } from '../src/defaultSettings';
+
+describe('default settings loading', () => {
+  beforeEach(() => {
+    localStorage.clear();
+  });
+
+  test('returns defaults when no settings stored', async () => {
+    const res = await storage.getItem('settings');
+    expect(res).toEqual({ settings: { ...defaultSettings } });
+    expect(getItemSync('settings')).toEqual({ settings: { ...defaultSettings } });
+  });
+});
